src/BionetDataCollection.Surveys.Web/Controllers/SurveyController.cs
// Copyright (c) Microsoft. All rights reserved.
// Licensed under the MIT license. See LICENSE file in the project root for full license information.

using System.Security.Claims;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.Extensions.Logging;
using BionetDataCollection.Surveys.Data.DataModels;
using BionetDataCollection.Surveys.Data.DTOs;
using BionetDataCollection.Surveys.Security.Policy;
using BionetDataCollection.Surveys.Web.Logging;
using BionetDataCollection.Surveys.Web.Models;
using BionetDataCollection.Surveys.Web.Services;
using System;
using System.Linq;
using Microsoft.Identity.Web;
using System.Net.Http;

namespace BionetDataCollection.Surveys.Web.Controllers
{
    /// <summary>
    /// This MVC controller provides actions for the management of <see cref="Survey"/>s.
    /// Most of the actions in this controller class require the user to be signed in.
    /// </summary>
    [AuthorizeForScopes(ScopeKeySection = "SurveyApi:Scope")]
    public class SurveyController : Controller
    {
        private readonly ISurveyService _surveyService;
        private readonly ILogger _logger;
        private readonly IAuthorizationService _authorizationService;

        public SurveyController(ISurveyService surveyService,
                                ILogger<SurveyController> logger,
                                IAuthorizationService authorizationService)
        {
            _surveyService = surveyService;
            _logger = logger;
            _authorizationService = authorizationService;
        }

        /// <summary>
        /// This action shows a list of <see cref="Survey"/>s related to the user. This includes <see cref="Survey"/>s that the user owns, 
        /// <see cref="Survey"/>s that the user contributes to, and <see cref="Survey"/>s the user has published.
        /// 
        /// This action also calls the <see cref="SurveyService"/> to process pending contributor requests.
        /// </summary>
        /// <returns>A view that shows the user's <see cref="Survey"/>s</returns>
        public async Task<IActionResult> Index()
        {
            try
            {
                // If there are any pending contributor requests that 
                await _surveyService.ProcessPendingContributorRequestsAsync();

                var userId = User.GetSurveyUserIdValue();
                var user = User.GetObjectIdentifierValue();
                var issuerValue = User.GetIssuerValue();
                var actionName = $"{typeof(SurveyController).FullName}.{nameof(Index)}";
                _logger.GetSurveysForUserOperationStarted(actionName, user, issuerValue);

                // The SurveyService.GetSurveysForUserAsync returns a UserSurveysDTO that has properties for Published, Own, and Contribute
                var result = await _surveyService.GetSurveysForUserAsync(userId);
                // If the user is in the creator role, the view shows a "Create Survey" button.
                var authResult = await _authorizationService.AuthorizeAsync(User, PolicyNames.RequireSurveyCreator);
                ViewBag.IsUserCreator = authResult?.Succeeded;
                _logger.GetSurveysForUserOperationSucceeded(actionName, user, issuerValue);
                return View(result);
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "403 Forbidden".Equals(requestEx.Message?.Trim()))
            {
                return ForbidenAccessToTheSurvey();
            }
            catch
            {
                ViewBag.Message = "Unexpected Error";
            }

            return View("~/Views/Shared/Error.cshtml");
        }

        /// <summary>
        /// This action shows a list of <see cref="Survey"/>s owned by users in the same <see cref="Tenant"/> as the current user.
        /// </summary>
        /// <returns>A view that shows <see cref="Survey"/>s in the same <see cref="Tenant"/> as the current user</returns>
        public async Task<IActionResult> ListPerTenant()
        {
            try
            {
                var tenantId = User.GetSurveyTenantIdValue();
                var surveys = await _surveyService.GetSurveysForTenantAsync(tenantId);
                // If the user is an administrator, additional functionality is exposed. 
                var authResult = await _authorizationService.AuthorizeAsync(User, PolicyNames.RequireSurveyAdmin);
                ViewBag.IsUserAdmin = authResult?.Succeeded;
                return View(surveys);
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "403 Forbidden".Equals(requestEx.Message?.Trim()))
            {
                return ForbidenAccessToTheSurvey();
            }
            catch
            {
                ViewBag.Message = "Unexpected Error";
            }
            return View("~/Views/Shared/Error.cshtml");
        }

        /// <summary>
        /// This action provides the Http Get experience for creating a <see cref="Survey"/>.
        /// This action is restricted to users in the survey creator role. 
        /// Creator role inclusion is implemented using the RequireSurveyCreator policy
        /// which is defined in <see cref="SurveyCreatorRequirement"/>.
        /// </summary>
        /// <returns>A view with form fields ued to create a <see cref="Survey"/></returns>
        [Authorize(Policy = PolicyNames.RequireSurveyCreator)]
        public IActionResult Create()
        {
            var survey = new SurveyDTO();
            return View(survey);
        }

        /// <summary>
        /// This action provides the Http Post experience for creating a <see cref="Survey"/>.
        /// This action is restricted to users in the survey creator role.
        /// </summary>
        /// <param name="survey">The <see cref="SurveyDTO"/> instance that contains the fields necessary to create a <see cref="Survey"/></param>
        /// <returns>A view that either shows validation errors or a redirection to the Survey Edit experience</returns>
        [HttpPost]
        [Authorize(Policy = PolicyNames.RequireSurveyCreator)]
        [ValidateAntiForgeryToken]
        public async Task<IActionResult> Create(SurveyDTO survey)
        {
            try
            {
                if (ModelState.IsValid)
                {
                    var result = await _surveyService.CreateSurveyAsync(survey);
                    return RedirectToAction("Edit", new { id = result.Id });
                }
                else
                {
                    ModelState.AddModelError(string.Empty, "Bad Request");
                    return View(survey);
                }
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "403 Forbidden".Equals(requestEx.Message?.Trim()))
            {
                ForbidenAccessToTheSurvey(survey);
            }
            catch
            {
                ModelState.AddModelError(string.Empty, "Unable to create survey.");
            }
            return View("~/Views/Shared/Error.cshtml");
        }

        /// <summary>
        /// This action shows the details of a specific <see cref="Survey"/>.
        /// </summary>
        /// <param name="id">The id of the <see cref="Survey"/></param>
        /// <returns>A view showing the contents of a <see cref="Survey"/>, or an error message if the <see cref="Survey"/> is not found</returns>
        public async Task<IActionResult> Details(int id)
        {
            try
            {
                var survey = await _surveyService.GetSurveyAsync(id);
                return View(survey);
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "403 Forbidden".Equals(requestEx.Message?.Trim()))
            {
                return ForbidenAccessToTheSurvey();
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "404 NotFound".Equals(requestEx.Message?.Trim()))
            {
                ModelState.AddModelError(string.Empty, "The survey can not be found");
                ViewBag.Message = "The survey can not be found";
                return View("~/Views/Shared/Error.cshtml");
            }
            catch
            {
                ViewBag.Message = "Unexpected Error";
            }
            return View("~/Views/Shared/Error.cshtml");
        }

        /// <summary>
        /// This action provides the Http Get experience for editing a <see cref="Survey"/>.
        /// </summary>
        /// <param name="id">The id of the <see cref="Survey"/></param>
        /// <returns>A view showing the <see cref="Survey"/>'s title and questions</returns>
        public async Task<IActionResult> Edit(int id)
        {
            try
            {
                var survey = await _surveyService.GetSurveyAsync(id);

                if (survey.Published)
                {
                    ViewBag.Message = "The survey is already published! You need to unpublish it in order to edit.";
                    return View("~/Views/Shared/Error.cshtml");
                }

                return View(survey);
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "403 Forbidden".Equals(requestEx.Message?.Trim()))
            {
                return ForbidenAccessToTheSurvey();
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "404 NotFound".Equals(requestEx.Message?.Trim()))
            {
                return SurveyNotFound();
            }
            catch
            {
                ViewBag.Message = "Unexpected Error";
            }
            return View("~/Views/Shared/Error.cshtml");
        }

        /// <summary>
        /// This action provides the Http Get experience for editing the title of a <see cref="Survey"/>
        /// </summary>
        /// <param name="id">The id of the <see cref="Survey"/></param>
        /// <returns>A view with form fields for the <see cref="Survey"/> being edited</returns>
        public async Task<IActionResult> EditTitle(int id)
        {
            try
            {
                var survey = await _surveyService.GetSurveyAsync(id);
                var model = survey;
                model.ExistingTitle = model.Title;
                return View(model);
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "403 Forbidden".Equals(requestEx.Message?.Trim()))
            {
                return ForbidenAccessToTheSurvey();
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "404 NotFound".Equals(requestEx.Message?.Trim()))
            {
                return SurveyNotFound();
            }
            catch
            {
                ViewBag.Message = "Unexpected Error";
            }
            return View("~/Views/Shared/Error.cshtml");
        }

        /// <summary>
        /// This action provides the Http Post experience for editing a <see cref="Survey"/>.
        /// </summary>
        /// <param name="model">The <see cref="SurveyDTO"/> instance that contains the <see cref="Survey"/>'s updated fields</param>
        /// <returns>A view that either shows validation errors or a redirection to the Survey Edit experience</returns>
        [HttpPost]
        [ValidateAntiForgeryToken]
        // We want to bind Id and Title and exclude Published, we don't want to publish when editing
        public async Task<IActionResult> EditTitle([Bind("Id", "Title", "ExistingTitle")] SurveyDTO model)
        {
            try
            {
                if (ModelState.IsValid)
                {
                    var result = await _surveyService.UpdateSurveyAsync(model);
                    return RedirectToAction("Edit", new { id = model.Id });
                }
                else
                {
                    ModelState.AddModelError(string.Empty, "Bad Request");
                    return View("EditTitle", model);
                }
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "403 Forbidden".Equals(requestEx.Message?.Trim()))
            {
                ForbidenAccessToTheSurvey(model);
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "404 NotFound".Equals(requestEx.Message?.Trim()))
            {
                return SurveyNotFound();
            }
            catch
            {
                ModelState.AddModelError(string.Empty, "Unable to save changes.");
            }
            return View("~/Views/Shared/Error.cshtml");
        }

        /// <summary>
        /// This action provides the Http Get experience for deleting a <see cref="Survey"/>.
        /// </summary>
        /// <param name="id">The id for the <see cref="Survey"/></param>
        /// <returns>A view that shows a delete confirmation prompt</returns>
        public async Task<IActionResult> Delete(int id)
        {
            try
            {
                var survey = await _surveyService.GetSurveyAsync(id);
                return View(survey);
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "403 Forbidden".Equals(requestEx.Message?.Trim()))
            {
                return ForbidenAccessToTheSurvey();
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "404 NotFound".Equals(requestEx.Message?.Trim()))
            {
                return SurveyNotFound();
            }
            catch
            {
                ViewBag.Message = "Unexpected Error";
            }
            return View("~/Views/Shared/Error.cshtml");
        }

        /// <summary>
        /// This action provides the Http Post experience for deleting a <see cref="Survey"/>.
        /// </summary>
        /// <param name="model">The <see cref="SurveyDTO"/> instance that contains the id of the <see cref="Survey"/> to be deleted</param>
        /// <returns>A view that either shows validation errors or a redirection to the Survey Edit experience</returns>
        [HttpPost]
        [ValidateAntiForgeryToken]
        public async Task<IActionResult> Delete([Bind("Id")] SurveyDTO model)
        {
            try
            {
                var surveyResult = await _surveyService.GetSurveyAsync(model.Id);
                var result = await _surveyService.DeleteSurveyAsync(model.Id);
                ViewBag.Message = "The following survey has been deleted.";
                return View("DeleteResult", result);
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "403 Forbidden".Equals(requestEx.Message?.Trim()))
            {
                return ForbidenAccessToTheSurvey(model);
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "404 NotFound".Equals(requestEx.Message?.Trim()))
            {
                return SurveyNotFound();
            }
            catch
            {
                ViewBag.Message = "Unexpected Error";
            }
            return View("~/Views/Shared/Error.cshtml");
        }

        /// <summary>
        /// This action shows a list of contributors associated with a specific <see cref="Survey"/>.
        /// </summary>
        /// <param name="id">The id of the <see cref="Survey"/></param>
        /// <returns>A view showing contributors associated with a <see cref="Survey"/></returns>
        public async Task<IActionResult> Contributors(int id)
        {
            try
            {
                var result = await _surveyService.GetSurveyContributorsAsync(id);
                return View(result);
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "403 Forbidden".Equals(requestEx.Message?.Trim()))
            {
                return ForbidenAccessToTheSurvey();
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "404 NotFound".Equals(requestEx.Message?.Trim()))
            {
                return SurveyNotFound();
            }
            catch
            {
                ViewBag.Message = "Unexpected Error";
            }
            return View("~/Views/Shared/Error.cshtml");
        }

        /// <summary>
        /// This action provides the Http Get experience for creating a <see cref="ContributorRequest"/>.
        /// </summary>
        /// <param name="id">The id of the <see cref="Survey"/></param>
        /// <returns>A view with form fields used to create the <see cref="ContributorRequest"/></returns>
        public async Task<IActionResult> RequestContributor(int id)
        {
            try
            {
                var survey = await _surveyService.GetSurveyAsync(id);
                ViewBag.SurveyId = id;
                return View();
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "403 Forbidden".Equals(requestEx.Message?.Trim()))
            {
                return ForbidenAccessToTheSurvey();
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "404 NotFound".Equals(requestEx.Message?.Trim()))
            {
                return SurveyNotFound();
            }
            catch
            {
                ViewBag.Message = "Unexpected Error";
            }
            return View("~/Views/Shared/Error.cshtml");
        }

        /// <summary>
        /// This action provides the Http Post experience for creating a <see cref="ContributorRequest"/>.
        /// </summary>
        /// <param name="contributorRequestViewModel">The <see cref="SurveyContributorRequestViewModel"/> instance with fields used to create a new <see cref="ContributorRequest"/></param>
        /// <returns>A redirection to the Show Contributors experience if persistance succeeds, or a view showing validation errors if not.</returns>
        [HttpPost]
        [ValidateAntiForgeryToken]
        public async Task<IActionResult> RequestContributor(SurveyContributorRequestViewModel contributorRequestViewModel)
        {
            try
            {
                if (!ModelState.IsValid)
                {
                    return RedirectToAction(nameof(Contributors), new { id = contributorRequestViewModel.SurveyId });
                }

                var existingContributors = await _surveyService.GetSurveyContributorsAsync(contributorRequestViewModel.SurveyId);
                if (existingContributors.Contributors.Any(item =>
                         String.Equals(item.Email, contributorRequestViewModel.EmailAddress, StringComparison.OrdinalIgnoreCase)))
                {
                    ViewBag.SurveyId = contributorRequestViewModel.SurveyId;
                    ViewBag.Message = contributorRequestViewModel.EmailAddress + " is already a contributor";
                    return View();
                }

                if (existingContributors.Requests.Any(item =>
                    String.Equals(item.EmailAddress, contributorRequestViewModel.EmailAddress, StringComparison.OrdinalIgnoreCase)))
                {
                    ViewBag.SurveyId = contributorRequestViewModel.SurveyId;
                    ViewBag.Message = contributorRequestViewModel.EmailAddress + " has already been requested before";
                    return View();
                }

                await _surveyService.AddContributorRequestAsync(new ContributorRequest
                {
                    SurveyId = contributorRequestViewModel.SurveyId,
                    EmailAddress = contributorRequestViewModel.EmailAddress
                });

                ViewBag.Message = $"Contribution Requested for {contributorRequestViewModel.EmailAddress}";
                ViewBag.SurveyId = contributorRequestViewModel.SurveyId;
                var result = await _surveyService.GetSurveyContributorsAsync(contributorRequestViewModel.SurveyId);
                return View("Contributors", result);
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "403 Forbidden".Equals(requestEx.Message?.Trim()))
            {
                return ForbidenAccessToTheSurvey();
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "404 NotFound".Equals(requestEx.Message?.Trim()))
            {
                return SurveyNotFound();
            }
            catch
            {
                ViewBag.Message = "Unexpected Error";
            }
            return View("~/Views/Shared/Error.cshtml");
        }

        /// <summary>
        /// This action provides the Http Get experience for publishing a <see cref="Survey"/>.
        /// </summary>
        /// <param name="id">The id of the <see cref="Survey"/></param>
        /// <returns>A page that asks the user to confirm that he/she wants to publish this <see cref="Survey"/></returns>
        public async Task<IActionResult> Publish(int id)
        {
            try
            {
                var survey = await _surveyService.GetSurveyAsync(id);
                if (survey.Published)
                {
                    ModelState.AddModelError(string.Empty, $"The survey is already published");
                    return View("PublishResult", survey);
                }
                else
                {
                    return View(survey);
                }
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "403 Forbidden".Equals(requestEx.Message?.Trim()))
            {
                return ForbidenAccessToTheSurvey();
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "404 NotFound".Equals(requestEx.Message?.Trim()))
            {
                return SurveyNotFound();
            }
            catch
            {
                ViewBag.Message = "Unexpected Error";
            }
            return View("~/Views/Shared/Error.cshtml");
        }

        /// <summary>
        /// This action provides the Http Post experience for publishing a <see cref="Survey"/>.
        /// </summary>
        /// <param name="model">The <see cref="SurveyDTO"/> instance that has the id field used to publish a <see cref="Survey"/></param>
        /// <returns>A confirmation page showing that the <see cref="Survey"/> was published, or errors</returns>
        [HttpPost]
        [ValidateAntiForgeryToken]
        public async Task<IActionResult> Publish([Bind("Id")] SurveyDTO model)
        {
            try
            {
                var surveyResult = await _surveyService.GetSurveyAsync(model.Id);
                if (surveyResult.Published)
                {
                    ModelState.AddModelError(string.Empty, $"The survey is already published");
                    return View("PublishResult", surveyResult);
                }
                else
                {
                    var publishResult = await _surveyService.PublishSurveyAsync(model.Id);
                    ViewBag.Message = "The following survey has been published.";
                    return View("PublishResult", publishResult);
                }
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "403 Forbidden".Equals(requestEx.Message?.Trim()))
            {
                return ForbidenAccessToTheSurvey();
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "404 NotFound".Equals(requestEx.Message?.Trim()))
            {
                return SurveyNotFound();
            }
            catch
            {
                ViewBag.Message = "Unexpected Error";
            }
            return View("~/Views/Shared/Error.cshtml");
        }

        /// <summary>
        /// This action provides the Http Get experience for unpublishing a <see cref="Survey"/>.
        /// </summary>
        /// <param name="id">The id of the <see cref="Survey"/></param>
        /// <returns>A page that asks the user to confirm that he/she wants to publish this <see cref="Survey"/></returns>
        public async Task<IActionResult> UnPublish(int id)
        {
            try
            {
                var survey = await _surveyService.GetSurveyAsync(id);
                if (!survey.Published)
                {
                    ModelState.AddModelError(string.Empty, $"The survey is already unpublished");
                    return View("UnPublishResult", survey);
                }
                else
                {
                    return View(survey);
                }
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "403 Forbidden".Equals(requestEx.Message?.Trim()))
            {
                return ForbidenAccessToTheSurvey();
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "404 NotFound".Equals(requestEx.Message?.Trim()))
            {
                return SurveyNotFound();
            }
            catch
            {
                ViewBag.Message = "Unexpected Error";
            }
            return View("~/Views/Shared/Error.cshtml");
        }

        /// <summary>
        /// This action provides the Http Post experience for unpublishing a <see cref="Survey"/>.
        /// </summary>
        /// <param name="model">The <see cref="SurveyDTO"/> instance that has the id field used to unpublish a <see cref="Survey"/></param>
        /// <returns>A confirmation page showing that the <see cref="Survey"/> was unpublished, or errors</returns>
        [HttpPost]
        [ValidateAntiForgeryToken]
        public async Task<IActionResult> UnPublish([Bind("Id")] SurveyDTO model)
        {
            try
            {
                var surveyResult = await _surveyService.GetSurveyAsync(model.Id);
                if (!surveyResult.Published)
                {
                    ModelState.AddModelError(string.Empty, $"The survey is already unpublished");
                    return View("UnPublishResult", surveyResult);
                }
                else
                {
                    var unpublishResult = await _surveyService.UnPublishSurveyAsync(model.Id);
                    ViewBag.Message = "The following survey has been unpublished.";
                    return View("UnPublishResult", unpublishResult);
                }
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "403 Forbidden".Equals(requestEx.Message?.Trim()))
            {
                return ForbidenAccessToTheSurvey();
            }
            catch (HttpRequestException requestEx) when ("Microsoft.Identity.Web".Equals(requestEx.Source) && "404 NotFound".Equals(requestEx.Message?.Trim()))
            {
                return SurveyNotFound();
            }
            catch
            {
                ViewBag.Message = "Unexpected Error";
            }
            return View("~/Views/Shared/Error.cshtml");
        }

        private IActionResult ForbidenAccessToTheSurvey()
        {
            ModelState.AddModelError(string.Empty, "Forbidden Access to the survey");
            ViewBag.Message = "Forbidden Access to the survey";
            return View("~/Views/Shared/Error.cshtml");
        }

        private IActionResult ForbidenAccessToTheSurvey(SurveyDTO survey)
        {
            ViewBag.Forbidden = true;
            return View(survey);
        }

        private IActionResult SurveyNotFound()
        {
            ModelState.AddModelError(string.Empty, "The survey can not be found");
            ViewBag.Message = "The survey can not be found";
            return View("~/Views/Shared/Error.cshtml");
        }
    }
}
